# go/users/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Users Service
    description: Service for managing users
    version: "1.0"
servers:
    - url: http://0.0.0.0:80/api/users
paths:
    /api/users:
        delete:
            tags:
                - users
            summary: DeleteUser users
            operationId: users#DeleteUser
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DeleteUserRequestBody'
                        example:
                            id: 2810239129742510560
            responses:
                "200":
                    description: OK response.
                "400":
                    description: 'BadRequest: Bad Request response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
                "404":
                    description: 'NotFound: Not Found response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
        post:
            tags:
                - users
            summary: CreateUser users
            operationId: users#CreateUser
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateUserRequestBody'
                        example:
                            display_name: Aspernatur quam repellat dolor laboriosam ut.
                            username: Quidem quis at.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                            example:
                                bio: Ad maiores aliquid neque accusamus tempore.
                                created_at: "1999-08-02T01:32:35Z"
                                display_name: Eum blanditiis facere unde dolor.
                                id: 279637141475456824
                                updated_at: "2011-11-02T19:28:26Z"
                                username: Error inventore accusamus.
                "400":
                    description: 'BadRequest: Bad Request response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
    /api/users/{id}:
        get:
            tags:
                - users
            summary: FindUserByID users
            operationId: users#FindUserByID
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    example: 4703302081872238694
                    format: int64
                  example: 6299968532845105888
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                            example:
                                bio: Minus eaque tempora in laborum.
                                created_at: "1993-06-18T08:42:53Z"
                                display_name: Praesentium similique nisi sit.
                                id: 4697244773142397205
                                updated_at: "1980-10-18T14:25:17Z"
                                username: Veniam iure id autem.
                "404":
                    description: 'NotFound: Not Found response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
    /api/users/{id}/bio:
        post:
            tags:
                - users
            summary: UpdateBio users
            operationId: users#UpdateBio
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    example: 1237231028843673176
                    format: int64
                  example: 3175780163654343168
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateBioRequestBody'
                        example:
                            bio: Voluptas nobis omnis.
            responses:
                "200":
                    description: OK response.
                "400":
                    description: 'BadRequest: Bad Request response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
                "404":
                    description: 'NotFound: Not Found response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
    /api/users/{id}/username:
        post:
            tags:
                - users
            summary: UpdateUsername users
            operationId: users#UpdateUsername
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    example: 395274392950160444
                    format: int64
                  example: 6966835270155161708
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateUsernameRequestBody'
                        example:
                            username: Odit et omnis placeat similique voluptatem id.
            responses:
                "200":
                    description: OK response.
                "400":
                    description: 'BadRequest: Bad Request response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
                "404":
                    description: 'NotFound: Not Found response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
    /api/users/follow:
        delete:
            tags:
                - users
            summary: Unfollow users
            operationId: users#Unfollow
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/FollowRequestBody'
                        example:
                            followee_id: 6202834779766302284
                            follower_id: 4698722160625222812
            responses:
                "200":
                    description: OK response.
                "400":
                    description: 'BadRequest: Bad Request response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
        post:
            tags:
                - users
            summary: Follow users
            operationId: users#Follow
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/FollowRequestBody'
                        example:
                            followee_id: 8492160358191620196
                            follower_id: 6881001417419968319
            responses:
                "200":
                    description: OK response.
                "400":
                    description: 'BadRequest: Bad Request response.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
    /swagger.json:
        get:
            tags:
                - users
            summary: Download ./gen/http/openapi.json
            operationId: users#/swagger.json
            responses:
                "200":
                    description: File downloaded
components:
    schemas:
        CreateUserRequestBody:
            type: object
            properties:
                display_name:
                    type: string
                    example: Saepe dolorem dolorum facere omnis.
                username:
                    type: string
                    example: Accusamus eum est ipsam placeat.
            example:
                display_name: Veritatis saepe sit soluta amet eos autem.
                username: Repellendus ipsum perferendis ipsum non.
            required:
                - username
                - display_name
        DeleteUserRequestBody:
            type: object
            properties:
                id:
                    type: integer
                    example: 5825090350270624949
                    format: int64
            example:
                id: 3392079364039272528
            required:
                - id
        Error:
            type: object
            properties:
                fault:
                    type: boolean
                    description: Is the error a server-side fault?
                    example: true
                id:
                    type: string
                    description: ID is a unique identifier for this particular occurrence of the problem.
                    example: 123abc
                message:
                    type: string
                    description: Message is a human-readable explanation specific to this occurrence of the problem.
                    example: parameter 'p' must be an integer
                name:
                    type: string
                    description: Name is the name of this class of errors.
                    example: bad_request
                temporary:
                    type: boolean
                    description: Is the error temporary?
                    example: true
                timeout:
                    type: boolean
                    description: Is the error a timeout?
                    example: false
            example:
                fault: true
                id: 123abc
                message: parameter 'p' must be an integer
                name: bad_request
                temporary: true
                timeout: false
            required:
                - name
                - id
                - message
                - temporary
                - timeout
                - fault
        FollowRequestBody:
            type: object
            properties:
                followee_id:
                    type: integer
                    example: 7062099314157344595
                    format: int64
                follower_id:
                    type: integer
                    example: 8316079895195226046
                    format: int64
            example:
                followee_id: 3701092369318947588
                follower_id: 8608922527124215505
            required:
                - follower_id
                - followee_id
        UpdateBioRequestBody:
            type: object
            properties:
                bio:
                    type: string
                    example: Itaque nulla eligendi inventore.
            example:
                bio: Ut numquam.
            required:
                - bio
        UpdateUsernameRequestBody:
            type: object
            properties:
                username:
                    type: string
                    example: Repellat praesentium enim est qui.
            example:
                username: Enim consequatur saepe adipisci.
            required:
                - username
        User:
            type: object
            properties:
                bio:
                    type: string
                    example: Tenetur autem rerum.
                created_at:
                    type: string
                    example: "1975-04-08T13:01:07Z"
                    format: date-time
                display_name:
                    type: string
                    example: Mollitia repellendus.
                id:
                    type: integer
                    example: 1345625246019500489
                    format: int64
                updated_at:
                    type: string
                    example: "1972-03-02T21:46:49Z"
                    format: date-time
                username:
                    type: string
                    example: Ea animi.
            example:
                bio: Magni perspiciatis distinctio neque esse at corporis.
                created_at: "1975-11-03T05:35:47Z"
                display_name: Mollitia modi harum qui omnis sapiente ad.
                id: 4941914734583634128
                updated_at: "1997-08-26T16:04:02Z"
                username: Tempora veritatis ratione.
            required:
                - id
                - username
                - display_name
                - bio
                - created_at
                - updated_at
tags:
    - name: users
      description: The users service performs operations on users information.